Add completed Jupyter notebook
This notebook contains the code for the following tasks:
- Perform a Histogram of Oriented Gradients (HOG) feature extraction on
a labeled training set of images and train a classifier Linear SVM
classifier
- Implement a sliding-window technique and use your trained classifier to
search for vehicles in images.
- Run your pipeline on a video stream (start with the test_video.mp4 and
later implement on full project_video.mp4) and create a heat map of
recurring detections frame by frame to reject outliers and follow
detected vehicles.
- Estimate a bounding box for vehicles detected.